refactore(ayako): app.go and server/
