+++ ayako/api/users.go
@@ -22,7 +22,7 @@ func (h *UsersHandlers) Get(c echo.Context) error {
 
 	mode := c.Param("mode")
 
-	user, err := h.Store().User().Get(ctx, uint(userId), mode)
+	user, err := h.Store.User().Get(ctx, uint(userId), mode)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusNotFound, "User not found")
 	}
