+++ ayako/middlewares/reqest_context/global.go
@@ -1,13 +1,13 @@
 package reqest_context
 
 import (
-	"context"
+	"github.com/deissh/rl/ayako/app"
 	myctx "github.com/deissh/rl/ayako/ctx"
 	"github.com/labstack/echo/v4"
 )
 
 // GlobalMiddleware create new request reqest_context with all information about caller
-func GlobalMiddleware(ctx context.Context) echo.MiddlewareFunc {
+func GlobalMiddleware(app *app.App) echo.MiddlewareFunc {
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
 		return func(c echo.Context) error {
 			req := c.Request()
@@ -15,7 +15,7 @@ func GlobalMiddleware(ctx context.Context) echo.MiddlewareFunc {
 			// context is growing very quickly
 			// due to the fact that it is used directly
 			// so create a copy
-			copyCtx := ctx
+			copyCtx := app.Context
 
 			if userId, ok := c.Get("current_user_id").(uint); ok {
 				copyCtx = myctx.Pipe(copyCtx, myctx.SetUserID(userId))
