+++ ayako/server/scheduled_task.go
@@ -1,4 +1,4 @@
-package base
+package server
 
 import (
 	"fmt"
